Refactor some controllers
