+++ src/Controller/Pia/EvaluationController.php
@@ -8,7 +8,7 @@
  * file that was distributed with this source code.
  */
 
-namespace PiaApi\Controller;
+namespace PiaApi\Controller\Pia;
 
 use Symfony\Component\HttpFoundation\Request;
 use FOS\RestBundle\View\View;
