+++ templates/Structure/manageStructures.html.twig
@@ -1,0 +1,121 @@
+{% extends 'pia/Layout/base.html.twig' %}
+
+{% import 'pia/Layout/_macros.html.twig' as macro %}
+
+{% block title %}PIA User Management{% endblock title %}
+
+{% block navbar %}
+    {% include 'pia/Layout/navbar.html.twig' with {'activeItem': 'structures'} %}
+{% endblock navbar %}
+
+{% block body %}
+
+    <div class="ui grid">
+
+        <div class="twelve wide column">
+
+            <div class="ui segment">
+
+                <h2>
+                    <a href="?limit=1000" class="ui tiny button" style="float:right;">Voir 1000</a>
+                    Liste des structures
+                </h2>
+
+                <table class="ui compact striped table">
+                    <thead>
+                        <tr>
+                            <th>#</th>
+                            <th>Nom</th>
+                            <th>Nombre d'utilisateurs</th>
+                            <th>Créé le</th>
+                            <th class="right aligned">
+                                <i class="cog icon"></i>
+                            </th>
+                        </tr>
+                    </thead>
+                    <tbody>
+
+                        {% for structure in structures %}
+                            <tr>
+                                <td>{{ structure.id }}</td>
+                                <td>{{ structure.name }}</td>
+                                <td>
+                                    <span class="ui label">{{ structure.users|length }}</span>
+                                </td>
+                                <td>
+                                    {{ structure.createdAt|date('d/m/Y')}}
+                                </td>
+                                <td class="right aligned">
+                                    <div class="ui small buttons">
+                                        <a class="ui green tiny icon button" title="Modifier la structure" data-modal="data-modal" href="{{ path('manage_structures_edit_structure', {'structureId':structure.id}) }}">
+                                            <i class="pencil icon"></i>
+                                        </a>
+                                        <a class="ui red tiny icon button" title="Supprimer la structure" data-modal="data-modal" href="{{ path('manage_structures_remove_structure', {'structureId':structure.id}) }}">
+                                            <i class="trash icon"></i>
+                                        </a>
+                                    </div>
+                                </td>
+                            </tr>
+                        {% endfor %}
+
+                    </tbody>
+                </table>
+
+                {{ pagerfanta(structures, 'semantic_ui') }}
+            </div>
+        </div>
+
+        <div class="four wide column">
+            <div class="ui segment">
+
+                <h2>Ajouter une structure</h2>
+
+                {{ render(controller('PiaApi\\Controller\\BackOffice\\StructureController:addStructureAction')) }}
+
+            </div>
+        </div>
+    </div>
+
+{% endblock %}
+
+{% block javascripts %}
+    {{ parent() }}
+
+    <script>
+        $(document).on('click', 'a[data-modal]', (e) => {
+            e.preventDefault();
+            e.stopPropagation();
+            e.stopImmediatePropagation();
+
+            let button = $(e.currentTarget);
+            let url = button.attr('href');
+
+            let modal = $('.ui.modal');
+            modal.hide();
+
+            $.ajax({
+                url: url,
+                success: (html) => {
+                    let modal = $('.ui.modal');
+
+                    modal.find('.header').html(button.attr('title'));
+
+                    modal
+                        .find('.content')
+                        .html(html);
+
+                    modal.find('.content .negative, .content .deny, .content .cancel').on('click', (e) => {
+                        $(e.currentTarget).closest('.ui.modal').modal('hide');
+                    });
+
+                    modal.find('.actions').remove();
+
+                    $('.ui.checkbox').checkbox();
+                    $('.ui.dropdown').dropdown();
+
+                    modal.modal('show');
+                }
+            })
+        });
+    </script>
+{% endblock javascripts %}
