+++ templates/User/manageUsers.html.twig
@@ -5,9 +5,7 @@
 {% block title %}PIA User Management{% endblock title %}
 
 {% block navbar %}
-    {% include 'pia/Layout/navbar.html.twig' with {
-        'activeItem' : 'users'
-    } %}
+    {% include 'pia/Layout/navbar.html.twig' with {'activeItem': 'users'} %}
 {% endblock navbar %}
 
 {% block body %}
@@ -87,12 +85,6 @@
                                     <div class="ui small buttons">
                                         <a class="ui green tiny icon button" data-modal="data-modal" title="Modifier l'utilisateur" href="{{ path('manage_users_edit_user', {'userId':user.id}) }}">
                                             <i class="pencil icon"></i>
-                                        </a>
-                                        <a class="ui orange tiny icon button" data-modal="data-modal" title="Envoyer l'email de réinitialisation de mot de passe" href="{{ path('manage_users_send_reset_password_email', {'userId':user.id}) }}">
-                                            <i class="icons">
-                                                <i class="envelope outline icon"></i>
-                                                <i class="bottom right corner plus icon"></i>
-                                            </i>
                                         </a>
                                         <a class="ui red tiny icon button" data-modal="data-modal" title="Supprimer l'utilisateur" href="{{ path('manage_users_remove_user', {'userId':user.id}) }}">
                                             <i class="trash icon"></i>
