+++ templates/pia/Application/manageApplications.html.twig
@@ -113,7 +113,7 @@
 
                 <h2>Ajouter une application</h2>
 
-                {{ render(controller('PiaApi\\Controller\\OauthController:addApplicationAction')) }}
+                {{ render(controller('PiaApi\\Controller\\BackOffice\\OauthController:addApplicationAction')) }}
 
             </div>
         </div>
