+++ templates/pia/Structure/manageStructures.html.twig
@@ -136,14 +136,14 @@
 
                 <h2>Ajouter une structure</h2>
 
-                {{ render(controller('PiaApi\\Controller\\StructureController:addStructureAction')) }}
+                {{ render(controller('PiaApi\\Controller\\BackOffice\\StructureController:addStructureAction')) }}
 
             </div>
             <div class="ui segment">
 
                 <h2>Ajouter un type de structure</h2>
 
-                {{ render(controller('PiaApi\\Controller\\StructureController:addStructureTypeAction')) }}
+                {{ render(controller('PiaApi\\Controller\\BackOffice\\StructureController:addStructureTypeAction')) }}
 
             </div>
         </div>
